+++ Changelog.md
@@ -14,7 +14,7 @@ npm install --save applause-cli
 -----
 
 
-## v1.6 (unreleased)
+## v1.6
  - Update development dependencies
  - Allow not specifying a path to an NPM package.json file
  - Add `.set_author()` method to `Program`